% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-conversations.R
\name{conversations_archive}
\alias{conversations_archive}
\title{Post conversations archive}
\usage{
conversations_archive(channel, token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{channel}{ID of conversation to archive}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded. Requires scope:
\code{conversations:write}}
}
\value{
\code{TRUE} (invisibly) for success.
}
\description{
Archives a conversation.
}
\keyword{internal}
