% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-conversations-conversations_history.R
\name{conversations_history}
\alias{conversations_history}
\alias{req_conversations_history}
\title{Get conversations history}
\usage{
conversations_history(
  channel,
  latest = lubridate::now(),
  oldest = 0,
  inclusive = TRUE,
  include_all_metadata = FALSE,
  per_req = 200L,
  max_reqs = Inf,
  max_tries_per_req = 3,
  token = Sys.getenv("SLACK_API_TOKEN")
)

req_conversations_history(
  channel,
  latest = lubridate::now(),
  oldest = 0,
  inclusive = TRUE,
  include_all_metadata = FALSE,
  per_req = 200L,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{channel}{(\code{character}) Conversation ID to fetch history for.}

\item{latest}{(\code{datetime} or \code{double}) End of time range of messages to
include in results.}

\item{oldest}{(\code{datetime} or \code{double}) Start of time range of messages to
include in results.}

\item{inclusive}{(\code{logical}) Include messages with \code{latest} or \code{oldest}
timestamp in results only when either timestamp is specified.}

\item{include_all_metadata}{(\code{logical}) Return all metadata associated with
this message.}

\item{per_req}{(\code{integer}) The maximum number of items to return. Fewer than
the requested number of items may be returned, even if the end of the list
hasn't been reached. Must be an integer under 1000.}

\item{max_reqs}{(\code{integer}) The maximum number of separate requests to
perform. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{max_tries_per_req}{(\code{integer}) The maximum number of times to attempt
each individual request. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
\code{conversations_history()}: A channel's messages as a tibble.

\code{req_conversations_history()}: A \code{httr2_request} request object to
fetch a conversation's history of messages and events.
}
\description{
Fetches a conversation's history of messages and events.
}
