% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-conversations.R
\name{conversations_info}
\alias{conversations_info}
\title{Get conversations info}
\usage{
conversations_info(
  channel,
  include_locale,
  include_num_members,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{channel}{Conversation ID to learn more about}

\item{include_locale}{Set this to \code{true} to receive the locale for this
conversation. Defaults to \code{false}}

\item{include_num_members}{Set to \code{true} to include the member count for the
specified conversation. Defaults to \code{false}}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded. Requires scope:
\code{conversations:read}}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Retrieve information about a conversation.
}
\keyword{internal}
