% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-conversations.R
\name{conversations_join}
\alias{conversations_join}
\title{Post conversations join}
\usage{
conversations_join(token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded. Requires scope:
\code{channels:write}}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Joins an existing conversation.
}
\keyword{internal}
