% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-conversations-conversations_members.R
\name{conversations_members}
\alias{conversations_members}
\alias{req_conversations_members}
\title{Get conversations history}
\usage{
conversations_members(
  channel,
  per_req = 200L,
  max_reqs = Inf,
  max_tries_per_req = 3,
  token = Sys.getenv("SLACK_API_TOKEN")
)

req_conversations_members(
  channel,
  per_req = 200L,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{channel}{(\code{character}) ID of the conversation to retrieve members for.}

\item{per_req}{(\code{integer}) The maximum number of items to return. Fewer than
the requested number of items may be returned, even if the end of the list
hasn't been reached. Must be an integer under 1000.}

\item{max_reqs}{(\code{integer}) The maximum number of separate requests to
perform. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{max_tries_per_req}{(\code{integer}) The maximum number of times to attempt
each individual request. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper

\code{conversations_members()}: A list of user IDs belonging to the
members in a conversation.

\code{req_conversations_members()}: A \code{httr2_request} request object to
retrieve members of a conversation.
}
\description{
Retrieve members of a conversation.
}
\keyword{internal}
