% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-param_renames.R
\name{.apply_slack_param_names}
\alias{.apply_slack_param_names}
\title{Restore Slack API parameter names}
\usage{
.apply_slack_param_names(x)
}
\arguments{
\item{x}{(\code{list} or \code{character}) A named object that might have parameters to
rename.}
}
\value{
\code{x} with parameters renamed.
}
\description{
Translate package parameter names to Slack API parameter names.
}
\keyword{internal}
