% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-api.R
\name{slack_api_test}
\alias{slack_api_test}
\title{Get api test}
\usage{
slack_api_test(error, foo, token = Sys.getenv("SLACK_TOKEN"))
}
\arguments{
\item{error}{Error response to return}

\item{foo}{example property to return}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Checks API calling code.
}
