% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-auth.R
\name{slack_auth_revoke}
\alias{slack_auth_revoke}
\title{Get auth revoke}
\usage{
slack_auth_revoke(test, token = Sys.getenv("SLACK_TOKEN"))
}
\arguments{
\item{test}{Setting this parameter to \code{1} triggers a \emph{testing mode} where the specified token will not actually be revoked.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Revokes a token.
}
