% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-auth.R
\name{slack_auth_test}
\alias{slack_auth_test}
\title{Get auth test}
\usage{
slack_auth_test(token = Sys.getenv("SLACK_TOKEN"))
}
\arguments{
\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Checks authentication & identity.
}
