% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-conversations.R
\name{slack_conversations_list}
\alias{slack_conversations_list}
\title{Get conversations list}
\usage{
slack_conversations_list(
  team_id = NULL,
  exclude_archived = FALSE,
  types = c("public_channel", "private_channel", "mpim", "im"),
  max_results = Inf,
  max_reqs = Inf,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{team_id}{(\code{character}) Encoded team id to list channels in, required if
token belongs to org-wide app.}

\item{exclude_archived}{(\code{logical}) Set to \code{TRUE} to exclude archived
channels from the list.}

\item{types}{(\code{character}) Mix and match channel types by providing a vector
of any combination of \code{public_channel}, \code{private_channel}, \code{mpim}, \code{im}.}
}
\value{
A tibble of information about channels and channel-like
conversations.
}
\description{
Lists all channels in a Slack team.
}
