% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversations-members.R
\name{slack_conversations_members}
\alias{slack_conversations_members}
\title{Get conversations members}
\usage{
slack_conversations_members(
  channel,
  max_results = Inf,
  max_reqs = Inf,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{max_results}{(\code{integer} or \code{Inf}) The maximum number of results to
return. Note that slightly more results may be returned if \code{max_results} is
not evenly divisible by 100.}

\item{max_reqs}{(\verb{length-1 integer}) The maximum number of separate requests
to perform. Passed to the max_reqs argument of
\code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}} when \code{next_req} is supplied. You will
mostly likely want to change the default value (\code{2}) to \code{Inf} after you
validate that the request works.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
A thread of messages posted to a conversation as a tibble. Note: The
parent message is always included in the response.
}
\description{
Retrieve members of a conversation.
}
\keyword{internal}
