% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-conversations.R
\name{slack_conversations_replies}
\alias{slack_conversations_replies}
\title{Get conversations replies}
\usage{
slack_conversations_replies(
  channel,
  ts,
  latest = lubridate::now(),
  oldest = 0,
  inclusive = TRUE,
  include_all_metadata = FALSE,
  max_results = Inf,
  max_reqs = Inf,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{channel}{(\code{character}) Conversation ID to fetch thread from.}

\item{ts}{(\code{character} or \code{datetime} or \code{double}) Unique identifier of either
a thread’s parent message or a message in the thread. ts must be the
timestamp of an existing message with 0 or more replies. If there are no
replies then just the single message referenced by ts will return - it is
just an ordinary, unthreaded message.}

\item{latest}{(\code{datetime} or \code{double}) Only messages before this Unix
timestamp will be included in results.}

\item{oldest}{(\code{datetime} or \code{double}) Only messages after this Unix
timestamp will be included in results.}

\item{inclusive}{(\code{logical}) Include messages with \code{latest} or \code{oldest}
timestamp in results only when either timestamp is specified.}

\item{include_all_metadata}{(\code{logical}) Return all metadata associated with
this message.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded. Requires scope:
\code{conversations:history}}
}
\value{
A thread of messages posted to a conversation as a tibble. Note: The
parent message is always included in the response.
}
\description{
Retrieve a thread of messages posted to a conversation
}
