% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030-pagination.R
\name{slack_pagination_cursor}
\alias{slack_pagination_cursor}
\title{Apply cursor-based pagination}
\usage{
slack_pagination_cursor(resp, req)
}
\arguments{
\item{resp}{(\code{httr2_response}) A \code{\link[httr2:response]{httr2::response()}} object (as returned by
\code{\link[httr2:req_perform]{httr2::req_perform()}}).}

\item{req}{(\code{httr2_request}) The request object to modify.}
}
\value{
A \code{\link[httr2:request]{httr2::request()}} object with additional class \code{nectar_request}.
}
\description{
Process a \code{resp} object to extract the next cursor, and build the next \code{req}.
Use as the \code{pagination_fn} argument to \code{\link[=slack_req_prepare]{slack_req_prepare()}}.
}
\keyword{internal}
