% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/930-responses.R
\name{slack_response_parser}
\alias{slack_response_parser}
\title{Basic response parsing}
\usage{
slack_response_parser(resp)
}
\arguments{
\item{resp}{A httr2 \link[httr2]{response} object, created by \code{\link[httr2:req_perform]{req_perform()}}.}
}
\value{
A tibble of results.
}
\description{
This function attempts to parse a Slack response into a tidy tibble. It will
gradually be replaced as I implement specific parsers, and eventually will
not be exported.
}
