% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users-info.R
\name{slack_users_info}
\alias{slack_users_info}
\title{Get users info}
\usage{
slack_users_info(
  user,
  include_locale = FALSE,
  max_results = Inf,
  max_reqs = Inf,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{user}{(\code{character}) User to get info on.}

\item{include_locale}{(\code{logical}) Set this to \code{TRUE} to receive the locale
for users. Defaults to \code{FALSE}.}

\item{max_results}{(\code{integer} or \code{Inf}) The maximum number of results to
return. Note that slightly more results may be returned if \code{max_results} is
not evenly divisible by 100.}

\item{max_reqs}{The maximum number of separate requests to perform. Passed to
the max_reqs argument of \code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}} when \code{next_req}
is supplied. The default \code{2} should likely be changed to \code{Inf} after you
validate the function.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
This method returns information about a member of a workspace.
}
\description{
Gets information about a user.
}
