% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users-list.R
\name{slack_users_list}
\alias{slack_users_list}
\title{Get users list}
\usage{
slack_users_list(
  include_locale = FALSE,
  team_id = NULL,
  max_results = Inf,
  max_reqs = Inf,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{include_locale}{(\code{logical}) Set this to \code{TRUE} to receive the locale
for users. Defaults to \code{FALSE}.}

\item{team_id}{(\code{character}) Encoded team id to list users in, required if
org token is used.}

\item{max_results}{(\code{integer} or \code{Inf}) The maximum number of results to
return. Note that slightly more results may be returned if \code{max_results} is
not evenly divisible by 100.}

\item{max_reqs}{The maximum number of separate requests to perform. Passed to
the max_reqs argument of \code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}} when \code{next_req}
is supplied. The default \code{2} should likely be changed to \code{Inf} after you
validate the function.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
This method returns a list of all users in the workspace. This
includes both invited users and deleted/deactivated users.
}
\description{
Lists all users in a Slack team.
}
