% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-users.R
\name{users_conversations}
\alias{users_conversations}
\title{Get users conversations}
\usage{
users_conversations(
  user,
  types,
  exclude_archived,
  limit,
  cursor,
  token = Sys.getenv("SLACK_TOKEN")
)
}
\arguments{
\item{user}{Browse conversations by a specific user ID's membership. Non-public channels are restricted to those where the calling user shares membership.}

\item{types}{Mix and match channel types by providing a comma-separated list of any combination of \code{public_channel}, \code{private_channel}, \code{mpim}, \code{im}}

\item{exclude_archived}{Set to \code{true} to exclude archived channels from the list}

\item{limit}{The maximum number of items to return. Fewer than the requested number of items may be returned, even if the end of the list hasn't been reached. Must be an integer no larger than 1000.}

\item{cursor}{Paginate through collections of data by setting the \code{cursor} parameter to a \code{next_cursor} attribute returned by a previous request's \code{response_metadata}. Default value fetches the first "page" of the collection. See \href{/docs/pagination}{pagination} for more detail.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
List conversations the calling user may access.
}
