% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-users.R
\name{users_get_presence}
\alias{users_get_presence}
\title{Get users get presence}
\usage{
users_get_presence(user, token = Sys.getenv("SLACK_TOKEN"))
}
\arguments{
\item{user}{User to get presence info on. Defaults to the authed user.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Gets user presence information.
}
