% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-users.R
\name{users_identity}
\alias{users_identity}
\title{Get users identity}
\usage{
users_identity(token = Sys.getenv("SLACK_TOKEN"))
}
\arguments{
\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Get a user's identity.
}
