% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-users.R
\name{users_lookup_by_email}
\alias{users_lookup_by_email}
\title{Get users lookup by email}
\usage{
users_lookup_by_email(email, token = Sys.getenv("SLACK_TOKEN"))
}
\arguments{
\item{email}{An email address belonging to a user in the workspace}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Find a user with an email address.
}
