% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-users.R
\name{users_set_presence}
\alias{users_set_presence}
\title{Post users set presence}
\usage{
users_set_presence(token = Sys.getenv("SLACK_TOKEN"))
}
\arguments{
\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Manually sets user presence.
}
