% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-conversations.R
\name{conversations_members}
\alias{conversations_members}
\title{Get conversations members}
\usage{
conversations_members(
  channel,
  max_results = Inf,
  max_reqs = Inf,
  cursor,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{channel}{ID of the conversation to retrieve members for}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded. Requires scope:
\code{conversations:read}}
}
\value{
BKTODO: Return descriptions are not yet implemented in beekeeper
}
\description{
Retrieve members of a conversation.
}
\keyword{internal}
