% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/910-call.R
\name{.slack_req_perform}
\alias{.slack_req_perform}
\title{Choose and apply pagination strategy}
\usage{
.slack_req_perform(req, pagination_fn = NULL, max_results, max_reqs, call)
}
\arguments{
\item{req}{(\code{httr2_request}) The request object to modify.}

\item{pagination_fn}{(\code{function}) The pagination scheme to use. Currently
either \code{NULL} (no pagination, the default) or \code{slack_pagination_cursor} to
use \code{\link[=slack_pagination_cursor]{slack_pagination_cursor()}} for \code{cursor}-based pagination. See
\href{https://api.slack.com/apis/pagination}{Pagination through collections} in
the Slack API documentation. We do not currently support "Classic
pagination".}

\item{max_results}{(\code{integer} or \code{Inf}) The maximum number of results to
return. Note that slightly more results may be returned if \code{max_results} is
not evenly divisible by 100.}

\item{max_reqs}{(\verb{length-1 integer}) The maximum number of separate requests
to perform. Passed to the max_reqs argument of
\code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}} when \code{next_req} is supplied. You will
mostly likely want to change the default value (\code{2}) to \code{Inf} after you
validate that the request works.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A list of \code{\link[httr2:response]{httr2::response()}} objects, one for each request
performed. The list has additional class \code{nectar_responses}.
}
\description{
Choose and apply pagination strategy
}
\keyword{internal}
