% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/910-call.R
\name{slack_call_api}
\alias{slack_call_api}
\title{Call the Slack Web API}
\usage{
slack_call_api(
  path,
  query = list(),
  body = NULL,
  method = NULL,
  pagination_fn = NULL,
  max_results = Inf,
  max_reqs = Inf,
  response_parser = slack_response_parser,
  token = Sys.getenv("SLACK_API_TOKEN"),
  call = rlang::caller_env()
)
}
\arguments{
\item{path}{(\code{character} or \code{list}) The route to an API endpoint. Optionally,
a list or character vector with the path as one or more unnamed arguments
(which will be concatenated with "/") plus named arguments to
\code{\link[glue:glue]{glue::glue()}} into the path.}

\item{query}{(\code{character} or \code{list}) An optional list or character vector of
parameters to pass in the query portion of the request. Can also include a
\code{.multi} argument to pass to \code{\link[httr2:req_url]{httr2::req_url_query()}} to control how
elements containing multiple values are handled.}

\item{body}{(multiple types) An object to use as the body of the request. If
any component of the body is a path, pass it through \code{\link[fs:path]{fs::path()}} or
otherwise give it the class "fs_path" to indicate that it is a path.}

\item{method}{(\verb{length-1 character}, optional) If the method is something
other than \code{GET} or \code{POST}, supply it. Case is ignored.}

\item{pagination_fn}{(\code{function}) The pagination scheme to use. Currently
either \code{NULL} (no pagination, the default) or \code{slack_pagination_cursor} to
use \code{\link[=slack_pagination_cursor]{slack_pagination_cursor()}} for \code{cursor}-based pagination. See
\href{https://api.slack.com/apis/pagination}{Pagination through collections} in
the Slack API documentation. We do not currently support "Classic
pagination".}

\item{max_results}{(\code{integer} or \code{Inf}) The maximum number of results to
return. Note that slightly more results may be returned if \code{max_results} is
not evenly divisible by 100.}

\item{max_reqs}{(\verb{length-1 integer}) The maximum number of separate requests
to perform. Passed to the max_reqs argument of
\code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}} when \code{next_req} is supplied. You will
mostly likely want to change the default value (\code{2}) to \code{Inf} after you
validate that the request works.}

\item{response_parser}{(\code{function}) A function to parse the server response.
Defaults to \code{\link[=slack_response_parser]{slack_response_parser()}}. Set this to \code{NULL} to return the raw
response from \code{\link[httr2:req_perform]{httr2::req_perform()}}.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}

\item{call}{(\code{environment}) The environment from which a function was called,
e.g. \code{\link[rlang:stack]{rlang::caller_env()}} (the default). The environment will be mentioned
in error messages as the source of the error. This argument is particularly
useful for functions that are intended to be called as utilities inside
other functions.}
}
\value{
A tibble with the results of the API call.
}
\description{
Generate and perform request to a Slack Web API method.
}
\keyword{internal}
