% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-conversations.R
\name{slack_conversations_history}
\alias{slack_conversations_history}
\title{Get conversations history}
\usage{
slack_conversations_history(
  channel,
  latest = lubridate::now(),
  oldest = 0,
  inclusive = TRUE,
  include_all_metadata = FALSE,
  max_results = Inf,
  max_reqs = Inf,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{channel}{(\code{character}) Conversation ID to fetch history for.}

\item{latest}{(\code{datetime} or \code{double}) End of time range of messages to
include in results.}

\item{oldest}{(\code{datetime} or \code{double}) Start of time range of messages to
include in results.}

\item{inclusive}{(\code{logical}) Include messages with \code{latest} or \code{oldest}
timestamp in results only when either timestamp is specified.}

\item{include_all_metadata}{(\code{logical}) Return all metadata associated with
this message.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded. Requires scope:
\code{conversations:history}}
}
\value{
A channel's messages as a tibble.
}
\description{
Fetches a conversation's history of messages and events.
}
