% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversations-list.R
\name{slack_conversations_list}
\alias{slack_conversations_list}
\title{Get conversations list}
\usage{
slack_conversations_list(
  team_id = NULL,
  exclude_archived = FALSE,
  types = c("public_channel", "private_channel", "mpim", "im"),
  max_results = Inf,
  max_reqs = Inf,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{team_id}{(\code{character}) Encoded team id to list channels in, required if
token belongs to org-wide app.}

\item{exclude_archived}{(\code{logical}) Set to \code{TRUE} to exclude archived
channels from the list.}

\item{types}{(\code{character}) Mix and match channel types by
providing a vector of any combination of \code{public_channel},
\code{private_channel}, \code{mpim}, \code{im}.}

\item{max_results}{(\code{integer} or \code{Inf}) The maximum number of results to
return. Note that slightly more results may be returned if \code{max_results} is
not evenly divisible by 100.}

\item{max_reqs}{The maximum number of separate requests to perform. Passed to
the max_reqs argument of \code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}} when \code{next_req}
is supplied. The default \code{2} should likely be changed to \code{Inf} after you
validate the function.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
A thread of messages posted to a conversation as a tibble. Note: The
parent message is always included in the response.
}
\description{
Lists all channels in a Slack team.
}
