% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversations-replies.R
\name{slack_conversations_replies}
\alias{slack_conversations_replies}
\title{Get conversation replies}
\usage{
slack_conversations_replies(
  channel,
  ts,
  latest = lubridate::now(),
  oldest = 0,
  inclusive = TRUE,
  include_all_metadata = FALSE,
  max_results = Inf,
  max_reqs = Inf,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{channel}{(\code{character}) Conversation ID to fetch thread from.}

\item{ts}{(\code{character} or \code{datetime} or \code{double}) Unique identifier of either
a thread’s parent message or a message in the thread. ts must be the
timestamp of an existing message with 0 or more replies. If there are no
replies then just the single message referenced by ts will return - it is
just an ordinary, unthreaded message.}

\item{latest}{(\code{datetime} or \code{double}) Only messages before this Unix
timestamp will be included in results.}

\item{oldest}{(\code{datetime} or \code{double}) Only messages after this Unix
timestamp will be included in results.}

\item{inclusive}{(\code{logical}) Include messages with \code{latest} or \code{oldest}
timestamp in results only when either timestamp is specified.}

\item{include_all_metadata}{(\code{logical}) Return all metadata associated with
this message.}

\item{max_results}{(\code{integer} or \code{Inf}) The maximum number of results to
return. Note that slightly more results may be returned if \code{max_results} is
not evenly divisible by 100.}

\item{max_reqs}{The maximum number of separate requests to perform. Passed to
the max_reqs argument of \code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}} when \code{next_req}
is supplied. The default \code{2} should likely be changed to \code{Inf} after you
validate the function.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
A thread of messages posted to a conversation as a tibble. Note: The
parent message is always included in the response.
}
\description{
Retrieve a thread of messages posted to a conversation.
}
