% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-prepare.R
\name{slack_req_prepare}
\alias{slack_req_prepare}
\title{Generate a request for the Slack API}
\usage{
slack_req_prepare(
  path,
  query = list(),
  body = NULL,
  method = NULL,
  pagination_fn = NULL,
  tidy_fn = nectar::resp_tidy_unknown,
  token = Sys.getenv("SLACK_API_TOKEN"),
  call = rlang::caller_env()
)
}
\arguments{
\item{path}{(\code{character} or \code{list}) The route to an API endpoint. Optionally,
a list or character vector with the path as one or more unnamed arguments
(which will be concatenated with "/") plus named arguments to
\code{\link[glue:glue]{glue::glue()}} into the path.}

\item{query}{(\code{character} or \code{list}) An optional list or character vector of
parameters to pass in the query portion of the request. Can also include a
\code{.multi} argument to pass to \code{\link[httr2:req_url]{httr2::req_url_query()}} to control how
elements containing multiple values are handled.}

\item{body}{(multiple types) An object to use as the body of the request. If
any component of the body is a path, pass it through \code{\link[fs:path]{fs::path()}} or
otherwise give it the class "fs_path" to indicate that it is a path.}

\item{method}{(\verb{length-1 character}, optional) If the method is something
other than \code{GET} or \code{POST}, supply it. Case is ignored.}

\item{pagination_fn}{(\code{function}) The pagination scheme to use. Currently
either \code{NULL} (no pagination, the default) or \code{slack_pagination_cursor} to
use \code{\link[=slack_pagination_cursor]{slack_pagination_cursor()}} for \code{cursor}-based pagination. See
\href{https://api.slack.com/apis/pagination}{Pagination through collections} in
the Slack API documentation. We do not currently support "Classic
pagination".}

\item{tidy_fn}{(\code{function}) A function that will be invoked by \code{\link[nectar:resp_tidy]{resp_tidy()}}
to tidy the response.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}

\item{call}{(\code{environment}) The environment from which a function was called,
e.g. \code{\link[rlang:stack]{rlang::caller_env()}} (the default). The environment will be mentioned
in error messages as the source of the error. This argument is particularly
useful for functions that are intended to be called as utilities inside
other functions.}
}
\value{
A \code{\link[httr2:request]{httr2::request()}} object with additional class \code{nectar_request}.
}
\description{
Prepare a request for the Slack API, using the opinionated framework defined
in \code{\link[nectar:req_init]{nectar::req_init()}}, \code{\link[nectar:req_modify]{nectar::req_modify()}}, \code{\link[nectar:req_tidy_policy]{nectar::req_tidy_policy()}},
and \code{\link[nectar:req_pagination_policy]{nectar::req_pagination_policy()}}.
}
\keyword{internal}
