% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackapi-package.R
\docType{package}
\name{slackapi-package}
\alias{slackapi}
\alias{slackapi-package}
\title{slackapi: `Slack Web API` Client}
\description{
A client to interact with the `Slack Web API` \url{https://api.slack.com/web}. The `Slack Web API` is an interface for querying information from and enacting change in a `Slack` workspace.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/slackapi}
  \item \url{https://jonthegeek.github.io/slackapi/}
  \item Report bugs at \url{https://github.com/jonthegeek/slackapi/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
