% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-users-users_list.R
\name{users_list}
\alias{users_list}
\alias{req_users_list}
\title{Get users list}
\usage{
users_list(
  include_locale = FALSE,
  team_id = NULL,
  per_req = 200L,
  max_reqs = Inf,
  max_tries_per_req = 3,
  token = Sys.getenv("SLACK_API_TOKEN")
)

req_users_list(
  include_locale = FALSE,
  team_id = NULL,
  per_req = 200L,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{include_locale}{(\code{logical}) Set this to \code{TRUE} to receive the locale
for users. Defaults to \code{FALSE}.}

\item{team_id}{(\code{character}) Encoded team id to list users in, required if
org token is used.}

\item{per_req}{(\code{integer}) The maximum number of items to return. Fewer than
the requested number of items may be returned, even if the end of the list
hasn't been reached. Must be an integer under 1000.}

\item{max_reqs}{(\code{integer}) The maximum number of separate requests to
perform. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{max_tries_per_req}{(\code{integer}) The maximum number of times to attempt
each individual request. Passed on to \code{\link[nectar:req_perform_opinionated]{nectar::req_perform_opinionated()}}.}

\item{token}{(\code{character}) A bearer token provided by Slack. A later
enhancement will add the ability to generate this token. Slack token are
long-lasting, and should be carefully guarded.}
}
\value{
\code{users_list()}: A list of all users in the workspace. This includes
both invited users and deleted/deactivated users.

\code{req_users_list()}: A \code{httr2_request} request object that lists all
users in a Slack team.
}
\description{
Lists all users in a Slack team.
}
